+++ src/qt/optionsmodel.cpp
@@ -163,9 +163,9 @@ QVariant OptionsModel::data(const QModelIndex & index, int role) const
         case ProxySocksVersion:
             return settings.value("nSocksVersion", 5);
         case Fee:
-            return QVariant((qint64) nTransactionFee);
+            return QVariant(nTransactionFee);
         case ReserveBalance:
-            return QVariant((qint64) nReserveBalance);
+            return QVariant(nReserveBalance);
         case DisplayUnit:
             return QVariant(nDisplayUnit);
         case DisplayAddresses:
@@ -244,12 +244,12 @@ bool OptionsModel::setData(const QModelIndex & index, const QVariant & value, in
         break;
         case Fee:
             nTransactionFee = value.toLongLong();
-            settings.setValue("nTransactionFee", (qint64) nTransactionFee);
+            settings.setValue("nTransactionFee", nTransactionFee);
             emit transactionFeeChanged(nTransactionFee);
             break;
         case ReserveBalance:
             nReserveBalance = value.toLongLong();
-            settings.setValue("nReserveBalance", (qint64) nReserveBalance);
+            settings.setValue("nReserveBalance", nReserveBalance);
             emit reserveBalanceChanged(nReserveBalance);
             break;
         case DisplayUnit:
